+++ src/test/java/io/keyko/nevermined/manager/AgreementsManagerIT.java
@@ -58,10 +58,13 @@ public class AgreementsManagerIT {
         metadataApiService = ManagerHelper.getMetadataService(config);
         String metadataUrl = config.getString("metadata-internal.url") + "/api/v1/metadata/assets/ddo/{did}";
         String provenanceUrl = config.getString("metadata-internal.url") + "/api/v1/metadata/assets/provenance/{did}";
-        String consumeUrl = config.getString("gateway.url") + "/api/v1/gateway/services/consume";
+        String gatewayUrl = config.getString("gateway.url");
+        String consumeUrl = gatewayUrl + "/api/v1/gateway/services/access";
         String secretStoreEndpoint = config.getString("secretstore.url");
         String providerAddress = config.getString("provider.address");
-        providerConfig = new ProviderConfig(consumeUrl, metadataUrl, provenanceUrl, secretStoreEndpoint, providerAddress);
+
+        providerConfig = new ProviderConfig(consumeUrl, metadataUrl, gatewayUrl, provenanceUrl, secretStoreEndpoint, providerAddress);
+
         neverminedAPI = NeverminedAPI.getInstance(config);
         Properties properties = new Properties();
         properties.put(NeverminedConfig.KEEPER_URL, config.getString("keeper.url"));